move commands into separate files
