+++ integrations/cypress/cypress.env.json
@@ -3,5 +3,6 @@
   "domain": "localhost:4400",
   "client_id": "YOUR_AUTH0_CLIENT_ID",
   "connection": "Username-Password-Authentication",
-  "scope": "openid profile email offline_access"
+  "scope": "openid profile email offline_access",
+  "auth0SessionCookieName": "appSession"
 }
